requirements.txt
PyJWT~=2.8.0
requests~=2.31.0
flask~=3.0.0